Reallocated the actual test files to follow PSR-0 principle. Added the usage of the bootstrap in phpunit.xml.dist and a filter for the coverage report. Also added tht GitList\Test namespace in the ./tests/bootstrap.php
