+++ tests/GitList/Test/Git/RepositoryTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace GitList\Git;
+namespace GitList\Test\Git;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator on 2012-09-22 at 20:17:20.
