+++ tests/GitList/Test/Provider/GitServiceProviderTest.php
@@ -1,11 +1,13 @@
 <?php
 
-namespace GitList\Provider;
+namespace GitList\Test\Provider;
+
+use GitList\Provider\GitServiceProvider;
 
 /**
- * Generated by PHPUnit_SkeletonGenerator on 2012-09-22 at 20:17:03.
+ * Generated by PHPUnit_SkeletonGenerator on 2012-09-22 at 20:17:15.
  */
-class ViewUtilServiceProviderTest
+class GitServiceProviderTest
     extends \PHPUnit_Framework_TestCase
 {
     /**
@@ -14,7 +16,7 @@ class ViewUtilServiceProviderTest
     protected static $app;
     
     /**
-     * @var ViewUtilServiceProvider
+     * @var GitServiceProvider
      */
     protected $object;
 
@@ -28,7 +30,7 @@ class ViewUtilServiceProviderTest
      * This method is called before a test is executed.
      */
     protected function setUp() {
-        $this->object = new ViewUtilServiceProvider;
+        $this->object = new GitServiceProvider;
     }
 
     /**
@@ -40,15 +42,15 @@ class ViewUtilServiceProviderTest
     }
 
     /**
-     * @covers GitList\Provider\ViewUtilServiceProvider::register
+     * @covers GitList\Provider\GitServiceProvider::register
      */
     public function testRegister() {
         $this->assertNull($this->object->register(self::$app));
-        $this->assertInstanceOf('GitList\Util\View', self::$app['util.view']);
+        $this->assertInstanceOf('GitList\Git\Client', self::$app['git']);
     }
 
     /**
-     * @covers GitList\Provider\ViewUtilServiceProvider::boot
+     * @covers GitList\Provider\GitServiceProvider::boot
      */
     public function testBoot() {
         $this->assertNull($this->object->boot(self::$app));
