+++ tests/GitList/Test/Provider/RepositoryUtilServiceProviderTest.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace GitList\Provider;
+namespace GitList\Test\Provider;
 
+use GitList\Provider\RepositoryUtilServiceProvider;
 /**
  * Generated by PHPUnit_SkeletonGenerator on 2012-09-22 at 20:17:09.
  */
