+++ tests/GitList/Test/Provider/ViewUtilServiceProviderTest.php
@@ -1,11 +1,13 @@
 <?php
 
-namespace GitList\Provider;
+namespace GitList\Test\Provider;
+
+use GitList\Provider\ViewUtilServiceProvider;
 
 /**
- * Generated by PHPUnit_SkeletonGenerator on 2012-09-22 at 20:17:15.
+ * Generated by PHPUnit_SkeletonGenerator on 2012-09-22 at 20:17:03.
  */
-class GitServiceProviderTest
+class ViewUtilServiceProviderTest
     extends \PHPUnit_Framework_TestCase
 {
     /**
@@ -14,7 +16,7 @@ class GitServiceProviderTest
     protected static $app;
     
     /**
-     * @var GitServiceProvider
+     * @var ViewUtilServiceProvider
      */
     protected $object;
 
@@ -28,7 +30,7 @@ class GitServiceProviderTest
      * This method is called before a test is executed.
      */
     protected function setUp() {
-        $this->object = new GitServiceProvider;
+        $this->object = new ViewUtilServiceProvider;
     }
 
     /**
@@ -40,15 +42,15 @@ class GitServiceProviderTest
     }
 
     /**
-     * @covers GitList\Provider\GitServiceProvider::register
+     * @covers GitList\Provider\ViewUtilServiceProvider::register
      */
     public function testRegister() {
         $this->assertNull($this->object->register(self::$app));
-        $this->assertInstanceOf('GitList\Git\Client', self::$app['git']);
+        $this->assertInstanceOf('GitList\Util\View', self::$app['util.view']);
     }
 
     /**
-     * @covers GitList\Provider\GitServiceProvider::boot
+     * @covers GitList\Provider\ViewUtilServiceProvider::boot
      */
     public function testBoot() {
         $this->assertNull($this->object->boot(self::$app));
