+++ tests/GitList/Test/Util/ViewTest.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace GitList\Util;
+namespace GitList\Test\Util;
+
+use GitList\Util\View;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator on 2012-09-22 at 20:14:05.
