+++ tests/bootstrap.php
@@ -1,2 +1,3 @@
 <?php
-    require __DIR__ . '/../vendor/autoload.php';
+    $loader = require __DIR__ . '/../vendor/autoload.php';
+    $loader->add('GitList\Test', __DIR__);
